refactor: export DebugInfo via dedicated file
