+++ frontend/src/features/prototype/debug-info/DebugInfo.ts
@@ -1,0 +1,1 @@
+export { default } from './components/organisms/DebugInfo';
